refactor: isolate web-specific test helpers to web workspace
